feat: Menu 생성 시 Request를 dto로 수정, MenuProducts 일급 컬렉션 생성

+++ src/main/java/kitchenpos/domain/menu/Menu.java
@@ -31,7 +31,7 @@ public class Menu extends BaseEntity {
         this.menuGroup = menuGroup;
     }
 
-    public Menu(String name, BigDecimal price, MenuGroup menuGroup) {
+    public Menu(String name, Long price, MenuGroup menuGroup) {
         this.name = name;
         this.price = new Price(price);
         this.menuGroup = menuGroup;
